requirements.txt
python3
numpy 
scipy
scikit-image
h5py
json
numpy-quaternion
matplotlib
tk
tensorboardx
torch
torchvision
utm
